revert custom filtering code
The upstream repo has solution for this now
But we still need this internal deployment until the upstream repo can support http 1.0.0
