+++ pubspec.yaml
@@ -2,25 +2,26 @@ name: instabug_http_client
 description:
   This package is an add on to instabug_flutter. It intercepts any requests performed
   with http Package and sends them to the report that will be sent to the dashboard.
-version: 3.1.0
+version: 3.2.0
 homepage: https://github.com/Instabug/Instabug-Flutter#readme
-publish_to: http://nexus.axlehire.com:7777
+publish_to: http://nexus.gojitsu.com:7777
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^1.2.1
-  instabug_flutter: '>=11.0.0 <14.0.0'
-  meta: ^1.12.0
-  watcher: ^1.1.0
+  http: ">=0.13.0 <2.0.0"
+  instabug_flutter: ">=11.0.0 <14.0.0"
+  meta: ^1.3.0
 
 dev_dependencies:
-  build_runner: ^2.4.11
-  flutter_lints: ^4.0.0
+  build_runner: ^2.1.1
   flutter_test:
     sdk: flutter
-  mockito: ^5.4.4
+  mockito: ^5.0.10
+# For information on the generic Dart part of this file, see the
+# following page: https://www.dartlang.org/tools/pub/pubspec
 
+# The following section is specific to Flutter.
