+++ test/instabug_http_client_test.mocks.dart
@@ -1,17 +1,15 @@
-// Mocks generated by Mockito 5.4.2 from annotations
+// Mocks generated by Mockito 5.4.0 from annotations
 // in instabug_http_client/test/instabug_http_client_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i7;
-import 'dart:convert' as _i8;
-import 'dart:typed_data' as _i9;
+import 'dart:async' as _i5;
+import 'dart:convert' as _i6;
+import 'dart:typed_data' as _i7;
 
-import 'package:http/http.dart' as _i4;
-import 'package:instabug_flutter/instabug_flutter.dart' as _i3;
-import 'package:instabug_http_client/src/instabug_http_client.dart' as _i6;
-import 'package:instabug_http_client/src/instabug_http_log_filter.dart' as _i2;
-import 'package:instabug_http_client/src/instabug_http_logger.dart' as _i5;
+import 'package:http/http.dart' as _i2;
+import 'package:instabug_http_client/instabug_http_client.dart' as _i4;
+import 'package:instabug_http_client/instabug_http_logger.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -25,9 +23,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeInstabugHttpLogFilter_0 extends _i1.SmartFake
-    implements _i2.InstabugHttpLogFilter {
-  _FakeInstabugHttpLogFilter_0(
+class _FakeClient_0 extends _i1.SmartFake implements _i2.Client {
+  _FakeClient_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -36,8 +33,9 @@ class _FakeInstabugHttpLogFilter_0 extends _i1.SmartFake
         );
 }
 
-class _FakeNetworkLogger_1 extends _i1.SmartFake implements _i3.NetworkLogger {
-  _FakeNetworkLogger_1(
+class _FakeInstabugHttpLogger_1 extends _i1.SmartFake
+    implements _i3.InstabugHttpLogger {
+  _FakeInstabugHttpLogger_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -46,8 +44,8 @@ class _FakeNetworkLogger_1 extends _i1.SmartFake implements _i3.NetworkLogger {
         );
 }
 
-class _FakeClient_2 extends _i1.SmartFake implements _i4.Client {
-  _FakeClient_2(
+class _FakeResponse_2 extends _i1.SmartFake implements _i2.Response {
+  _FakeResponse_2(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -56,30 +54,9 @@ class _FakeClient_2 extends _i1.SmartFake implements _i4.Client {
         );
 }
 
-class _FakeInstabugHttpLogger_3 extends _i1.SmartFake
-    implements _i5.InstabugHttpLogger {
-  _FakeInstabugHttpLogger_3(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-class _FakeResponse_4 extends _i1.SmartFake implements _i4.Response {
-  _FakeResponse_4(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-class _FakeStreamedResponse_5 extends _i1.SmartFake
-    implements _i4.StreamedResponse {
-  _FakeStreamedResponse_5(
+class _FakeStreamedResponse_3 extends _i1.SmartFake
+    implements _i2.StreamedResponse {
+  _FakeStreamedResponse_3(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -92,42 +69,15 @@ class _FakeStreamedResponse_5 extends _i1.SmartFake
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockInstabugHttpLogger extends _i1.Mock
-    implements _i5.InstabugHttpLogger {
+    implements _i3.InstabugHttpLogger {
   MockInstabugHttpLogger() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.InstabugHttpLogFilter get logFilter => (super.noSuchMethod(
-        Invocation.getter(#logFilter),
-        returnValue: _FakeInstabugHttpLogFilter_0(
-          this,
-          Invocation.getter(#logFilter),
-        ),
-      ) as _i2.InstabugHttpLogFilter);
-
-  @override
-  _i3.NetworkLogger get networkLogger => (super.noSuchMethod(
-        Invocation.getter(#networkLogger),
-        returnValue: _FakeNetworkLogger_1(
-          this,
-          Invocation.getter(#networkLogger),
-        ),
-      ) as _i3.NetworkLogger);
-
-  @override
-  set networkLogger(_i3.NetworkLogger? _networkLogger) => super.noSuchMethod(
-        Invocation.setter(
-          #networkLogger,
-          _networkLogger,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
   void onLogger(
-    _i4.Response? response, {
-    required DateTime? startTime,
+    _i2.Response? response, {
+    DateTime? startTime,
   }) =>
       super.noSuchMethod(
         Invocation.method(
@@ -143,65 +93,35 @@ class MockInstabugHttpLogger extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockInstabugHttpClient extends _i1.Mock
-    implements _i6.InstabugHttpClient {
+    implements _i4.InstabugHttpClient {
   MockInstabugHttpClient() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i4.Client get client => (super.noSuchMethod(
+  _i2.Client get client => (super.noSuchMethod(
         Invocation.getter(#client),
-        returnValue: _FakeClient_2(
+        returnValue: _FakeClient_0(
           this,
           Invocation.getter(#client),
         ),
-      ) as _i4.Client);
-
+      ) as _i2.Client);
   @override
-  _i5.InstabugHttpLogger get logger => (super.noSuchMethod(
+  _i3.InstabugHttpLogger get logger => (super.noSuchMethod(
         Invocation.getter(#logger),
-        returnValue: _FakeInstabugHttpLogger_3(
+        returnValue: _FakeInstabugHttpLogger_1(
           this,
           Invocation.getter(#logger),
         ),
-      ) as _i5.InstabugHttpLogger);
-
+      ) as _i3.InstabugHttpLogger);
   @override
-  set logger(_i5.InstabugHttpLogger? _logger) => super.noSuchMethod(
+  set logger(_i3.InstabugHttpLogger? _logger) => super.noSuchMethod(
         Invocation.setter(
           #logger,
           _logger,
         ),
         returnValueForMissingStub: null,
       );
-
-  @override
-  _i2.InstabugHttpLogFilter get logFilter => (super.noSuchMethod(
-        Invocation.getter(#logFilter),
-        returnValue: _FakeInstabugHttpLogFilter_0(
-          this,
-          Invocation.getter(#logFilter),
-        ),
-      ) as _i2.InstabugHttpLogFilter);
-
-  @override
-  _i3.NetworkLogger get networkLogger => (super.noSuchMethod(
-        Invocation.getter(#networkLogger),
-        returnValue: _FakeNetworkLogger_1(
-          this,
-          Invocation.getter(#networkLogger),
-        ),
-      ) as _i3.NetworkLogger);
-
-  @override
-  set networkLogger(_i3.NetworkLogger? _networkLogger) => super.noSuchMethod(
-        Invocation.setter(
-          #networkLogger,
-          _networkLogger,
-        ),
-        returnValueForMissingStub: null,
-      );
-
   @override
   void close() => super.noSuchMethod(
         Invocation.method(
@@ -210,13 +130,12 @@ class MockInstabugHttpClient extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
-  _i7.Future<_i4.Response> delete(
+  _i5.Future<_i2.Response> delete(
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i8.Encoding? encoding,
+    _i6.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -228,7 +147,7 @@ class MockInstabugHttpClient extends _i1.Mock
             #encoding: encoding,
           },
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #delete,
@@ -240,10 +159,9 @@ class MockInstabugHttpClient extends _i1.Mock
             },
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<_i4.Response> get(
+  _i5.Future<_i2.Response> get(
     Uri? url, {
     Map<String, String>? headers,
   }) =>
@@ -253,7 +171,7 @@ class MockInstabugHttpClient extends _i1.Mock
           [url],
           {#headers: headers},
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #get,
@@ -261,10 +179,9 @@ class MockInstabugHttpClient extends _i1.Mock
             {#headers: headers},
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<_i4.Response> head(
+  _i5.Future<_i2.Response> head(
     Uri? url, {
     Map<String, String>? headers,
   }) =>
@@ -274,7 +191,7 @@ class MockInstabugHttpClient extends _i1.Mock
           [url],
           {#headers: headers},
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #head,
@@ -282,14 +199,13 @@ class MockInstabugHttpClient extends _i1.Mock
             {#headers: headers},
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<_i4.Response> patch(
+  _i5.Future<_i2.Response> patch(
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i8.Encoding? encoding,
+    _i6.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -301,7 +217,7 @@ class MockInstabugHttpClient extends _i1.Mock
             #encoding: encoding,
           },
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #patch,
@@ -313,14 +229,13 @@ class MockInstabugHttpClient extends _i1.Mock
             },
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<_i4.Response> post(
+  _i5.Future<_i2.Response> post(
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i8.Encoding? encoding,
+    _i6.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -332,7 +247,7 @@ class MockInstabugHttpClient extends _i1.Mock
             #encoding: encoding,
           },
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #post,
@@ -344,14 +259,13 @@ class MockInstabugHttpClient extends _i1.Mock
             },
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<_i4.Response> put(
+  _i5.Future<_i2.Response> put(
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i8.Encoding? encoding,
+    _i6.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -363,7 +277,7 @@ class MockInstabugHttpClient extends _i1.Mock
             #encoding: encoding,
           },
         ),
-        returnValue: _i7.Future<_i4.Response>.value(_FakeResponse_4(
+        returnValue: _i5.Future<_i2.Response>.value(_FakeResponse_2(
           this,
           Invocation.method(
             #put,
@@ -375,10 +289,9 @@ class MockInstabugHttpClient extends _i1.Mock
             },
           ),
         )),
-      ) as _i7.Future<_i4.Response>);
-
+      ) as _i5.Future<_i2.Response>);
   @override
-  _i7.Future<String> read(
+  _i5.Future<String> read(
     Uri? url, {
     Map<String, String>? headers,
   }) =>
@@ -388,11 +301,10 @@ class MockInstabugHttpClient extends _i1.Mock
           [url],
           {#headers: headers},
         ),
-        returnValue: _i7.Future<String>.value(''),
-      ) as _i7.Future<String>);
-
+        returnValue: _i5.Future<String>.value(''),
+      ) as _i5.Future<String>);
   @override
-  _i7.Future<_i9.Uint8List> readBytes(
+  _i5.Future<_i7.Uint8List> readBytes(
     Uri? url, {
     Map<String, String>? headers,
   }) =>
@@ -402,30 +314,28 @@ class MockInstabugHttpClient extends _i1.Mock
           [url],
           {#headers: headers},
         ),
-        returnValue: _i7.Future<_i9.Uint8List>.value(_i9.Uint8List(0)),
-      ) as _i7.Future<_i9.Uint8List>);
-
+        returnValue: _i5.Future<_i7.Uint8List>.value(_i7.Uint8List(0)),
+      ) as _i5.Future<_i7.Uint8List>);
   @override
-  _i7.Future<_i4.StreamedResponse> send(_i4.BaseRequest? request) =>
+  _i5.Future<_i2.StreamedResponse> send(_i2.BaseRequest? request) =>
       (super.noSuchMethod(
         Invocation.method(
           #send,
           [request],
         ),
         returnValue:
-            _i7.Future<_i4.StreamedResponse>.value(_FakeStreamedResponse_5(
+            _i5.Future<_i2.StreamedResponse>.value(_FakeStreamedResponse_3(
           this,
           Invocation.method(
             #send,
             [request],
           ),
         )),
-      ) as _i7.Future<_i4.StreamedResponse>);
-
+      ) as _i5.Future<_i2.StreamedResponse>);
   @override
   void onLogger(
-    _i4.Response? response, {
-    required DateTime? startTime,
+    _i2.Response? response, {
+    DateTime? startTime,
   }) =>
       super.noSuchMethod(
         Invocation.method(
@@ -435,19 +345,4 @@ class MockInstabugHttpClient extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-}
-
-/// A class which mocks [NetworkLogger].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockNetworkLogger extends _i1.Mock implements _i3.NetworkLogger {
-  @override
-  _i7.Future<void> networkLog(_i3.NetworkData? data) => (super.noSuchMethod(
-        Invocation.method(
-          #networkLog,
-          [data],
-        ),
-        returnValue: _i7.Future<void>.value(),
-        returnValueForMissingStub: _i7.Future<void>.value(),
-      ) as _i7.Future<void>);
 }
