--- test/instabug_http_logger_test.mocks.dart
@@ -1,40 +1,0 @@
-// Mocks generated by Mockito 5.4.2 from annotations
-// in instabug_http_client/test/instabug_http_logger_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-
-import 'package:instabug_flutter/src/models/network_data.dart' as _i4;
-import 'package:instabug_flutter/src/modules/network_logger.dart' as _i2;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [NetworkLogger].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockNetworkLogger extends _i1.Mock implements _i2.NetworkLogger {
-  MockNetworkLogger() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i3.Future<void> networkLog(_i4.NetworkData? data) => (super.noSuchMethod(
-        Invocation.method(
-          #networkLog,
-          [data],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-}
